Feature: Mother Neonate ANC next visit date disable for Recovered patient status. 	- Implement swipe-to-refresh functionality on the referral history page and update the patient list with pagination changes. 	- MDT-14415

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/model/PatientsDataModel.kt
@@ -1,7 +1,7 @@
 package com.medtroniclabs.spice.model
 
 data class PatientsDataModel(
-    var skip: Int? = null,
+    var skip: Long? = null,
     var limit: Int? = null,
     val villageIds: List<Long>? = null,
     val searchText: String? = null,
